TransactionMemoAddressTextField refactored
