+++ VultisigApp/VultisigApp/Views/Components/TextFields/TransactionMemoAddressTextField.swift
@@ -10,10 +10,6 @@ import OSLog
 import UniformTypeIdentifiers
 import WalletCore
 
-#if os(iOS)
-import CodeScanner
-#endif
-
 struct TransactionMemoAddressTextField<MemoType: TransactionMemoAddressable>: View {
     @ObservedObject var memo: MemoType
     var addressKey: String
@@ -44,37 +40,7 @@ struct TransactionMemoAddressTextField<MemoType: TransactionMemoAddressable>: Vi
                 }
             }
             
-            content
-                .font(.body12Menlo)
-                .foregroundColor(.neutral0)
-                .frame(height: 48)
-                .frame(maxWidth: .infinity, alignment: .leading)
-                .background(Color.blue600)
-                .cornerRadius(10)
-#if os(iOS)
-                .sheet(isPresented: $showScanner) {
-                    codeScanner
-                }
-                .sheet(isPresented: $showImagePicker, onDismiss: processImage) {
-                    ImagePicker(selectedImage: $selectedImage)
-                }
-#elseif os(macOS)
-                .fileImporter(
-                    isPresented: $showImagePicker,
-                    allowedContentTypes: [UTType.image],
-                    allowsMultipleSelection: false
-                ) { result in
-                    do {
-                        let qrCode = try Utils.handleQrCodeFromImage(result: result)
-                        handleImageQrCode(data: qrCode)
-                    } catch {
-                        print(error)
-                    }
-                }
-                .onDrop(of: [.image], isTargeted: $isUploading) { providers -> Bool in
-                    OnDropQRUtils.handleOnDrop(providers: providers, handleImageQrCode: handleImageQrCode)
-                }
-#endif
+            container
         }
         .onChange(of: memo.addressFields[addressKey]) { oldValue, newValue in
             validateAddress(newValue ?? "")
@@ -108,53 +74,6 @@ struct TransactionMemoAddressTextField<MemoType: TransactionMemoAddressable>: Vi
                 .foregroundColor(.neutral0)
         }
     }
-    
-    var field: some View {
-        HStack(spacing: 0) {
-            TextField(addressKey.toFormattedTitleCase(), text: Binding<String>(
-                get: { memo.addressFields[addressKey] ?? "" },
-                set: { newValue in
-                    memo.addressFields[addressKey] = newValue
-                    DebounceHelper.shared.debounce {
-                        validateAddress(newValue)
-                    }
-                }
-            ))
-            .foregroundColor(.neutral0)
-            .submitLabel(.next)
-            .disableAutocorrection(true)
-            .borderlessTextFieldStyle()
-            .maxLength(Binding<String>(
-                get: { memo.addressFields[addressKey] ?? "" },
-                set: { newValue in
-                    memo.addressFields[addressKey] = newValue
-                    DebounceHelper.shared.debounce {
-                        validateAddress(newValue)
-                    }
-                }
-            ))
-#if os(iOS)
-            .textInputAutocapitalization(.never)
-            .keyboardType(.default)
-            .textContentType(.oneTimeCode)
-#endif
-            
-            pasteButton
-#if os(iOS)
-            scanButton
-#elseif os(macOS)
-            fileButton
-#endif
-            addressBookButton
-        }
-        .padding(.horizontal, 12)
-    }
-    
-#if os(iOS)
-    var codeScanner: some View {
-        QRCodeScannerView(showScanner: $showScanner, handleScan: handleScan)
-    }
-#endif
     
     var pasteButton: some View {
         Button {
@@ -208,57 +127,6 @@ struct TransactionMemoAddressTextField<MemoType: TransactionMemoAddressable>: Vi
         }
     }
     
-    private func processImage() {
-        guard let selectedImage = selectedImage else { return }
-        
-#if os(iOS)
-        handleImageQrCode(image: selectedImage)
-#endif
-    }
-    
-#if os(iOS)
-    private func handleScan(result: Result<ScanResult, ScanError>) {
-        switch result {
-        case .success(let result):
-            let qrCodeResult = result.string
-            memo.addressFields[addressKey] = qrCodeResult
-            validateAddress(memo.addressFields[addressKey] ?? "")
-            showScanner = false
-        case .failure(let err):
-            print("Failed to scan QR code, error: \(err.localizedDescription)")
-        }
-    }
-    
-    private func handleImageQrCode(image: UIImage) {
-        let qrCodeFromImage = Utils.handleQrCodeFromImage(image: image)
-        let address = String(data: qrCodeFromImage, encoding: .utf8) ?? ""
-        memo.addressFields[addressKey] = address
-        validateAddress(memo.addressFields[addressKey] ?? "")
-    }
-#elseif os(macOS)
-    private func handleImageQrCode(data: Data) {
-        let (address, amount, _) = Utils.parseCryptoURI(String(data: data, encoding: .utf8) ?? .empty)
-        memo.addressFields[addressKey] = address
-        memo.addressFields["amount"] = amount
-        validateAddress(address)
-    }
-#endif
-    
-    private func pasteAddress() {
-#if os(iOS)
-        if let clipboardContent = UIPasteboard.general.string {
-            memo.addressFields[addressKey] = clipboardContent
-            validateAddress(memo.addressFields[addressKey] ?? "")
-        }
-#elseif os(macOS)
-        let pasteboard = NSPasteboard.general
-        if let clipboardContent = pasteboard.string(forType: .string) {
-            memo.addressFields[addressKey] = clipboardContent
-            validateAddress(memo.addressFields[addressKey] ?? "")
-        }
-#endif
-    }
-    
     var optionalMessage: String {
         if isOptional {
             return " (optional)"
@@ -266,7 +134,7 @@ struct TransactionMemoAddressTextField<MemoType: TransactionMemoAddressable>: Vi
         return .empty
     }
     
-    private func validateAddress(_ newValue: String) {
+    func validateAddress(_ newValue: String) {
         
         if isOptional, newValue.isEmpty {
             isAddressValid = true
